massive update of unity and googleVR various bugfixes (mostly about raycast and vr) win-mouse build for promo materials purpose rebakes

+++ Assets/Scripts/GameController.cs
@@ -69,7 +69,7 @@ public class GameController
             return false;
         }
 
-        DataObjects.sceneInfo.tempScore += (int)(s * DataObjects.sceneInfo.multiplier + 0.5f);
+        DataObjects.sceneInfo.tempScore += (long)(s * DataObjects.sceneInfo.multiplier + 0.5f);
 
         TriggerOptionalScoreAction();
